Split router into multiple files
On the way, I did two more small changes:
- use #[serde(rename_all = "camelCase")] instead of manually renaming some fields
- the ping route now returns 204 (instead of 200), just as the node gateway does
