+++ src/router/ping_router.rs
@@ -1,0 +1,10 @@
+use rocket::{http::Status, Route};
+
+pub fn routes() -> Vec<Route> {
+    routes![ping]
+}
+
+#[get("/")]
+fn ping() -> Status {
+    Status::NoContent
+}
